requirements.txt
# Core Blogger App (blogger_post.py, scheduler.py)
google-api-python-client==2.100.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
schedule==1.2.0
python-dotenv==1.0.0
requests==2.32.5
openai>=1.0.0
google-generativeai>=0.3.0

# Realtime Data Collect Bot (bot.py)
feedparser>=6.0.0
trafilatura>=1.6.0



